pyproject.toml
[project]
name = "openai-agent-sdk-tools-calls-with-mongodb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dnspython>=2.7.0",
    "openai-agents>=0.2.0",
    "pymongo>=4.13.2",
    "python-dotenv>=1.1.1",
]
